#FDF-15: Merging Lean Memory changes to 3.0 branch TESTS_RAN: Unit test, dev test
git-svn-id: https://10.196.60.217/svn/schooner-products/branches/fdf-3.0@2431 01a69087-22d5-4a29-8152-8a1d5e10e5e9

+++ common/zstypes.h
@@ -35,7 +35,7 @@ typedef enum {
 } ZS_container_mode_t;
 
 
-typedef uint32_t baddr_t;
+typedef uint64_t baddr_t;
 typedef uint16_t cntr_id_t;
 typedef uint16_t hashsyn_t;
 typedef uint32_t ZS_time_t;
